+++ CHANGELOG.md
@@ -241,3 +241,4 @@
 * Define `->--` operator
 * Define `->>--` operator
 * Define `-->>--` operator
+* Rename `-/|/->` operator to `-/>/-`